# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_vmkzy_todoapp-go_internal_core_transport_http_response.ErrorResponse:
    properties:
      erorr:
        example: error
        type: string
      message:
        example: error message
        type: string
    type: object
  internal_features_statistics_transport_http.GetStatisticsResponse:
    properties:
      tasks_average_completion_time:
        example: 1m30s
        type: string
      tasks_completed:
        example: 5
        type: integer
      tasks_completed_rate:
        example: 50
        type: number
      tasks_created:
        example: 10
        type: integer
    type: object
  internal_features_tasks_transport_http.CreateTaskRequest:
    properties:
      author_user_id:
        example: 1
        type: integer
      description:
        example: сделать домашку
        maxLength: 1000
        minLength: 1
        type: string
      title:
        example: домашка
        maxLength: 100
        minLength: 1
        type: string
    required:
    - author_user_id
    - title
    type: object
  internal_features_tasks_transport_http.CreateTaskResponse:
    properties:
      author_user_id:
        example: 1
        type: integer
      completed:
        example: false
        type: boolean
      completed_at:
        example: "null"
        type: string
      created_at:
        example: "2026-08-26T10:30:00Z"
        type: string
      description:
        example: сделать домашку
        type: string
      id:
        example: 1
        type: integer
      title:
        example: домашка
        type: string
      version:
        example: 1
        type: integer
    type: object
  internal_features_tasks_transport_http.GetTaskResponse:
    properties:
      author_user_id:
        example: 1
        type: integer
      completed:
        example: false
        type: boolean
      completed_at:
        example: "null"
        type: string
      created_at:
        example: "2026-08-26T10:30:00Z"
        type: string
      description:
        example: сделать домашку
        type: string
      id:
        example: 1
        type: integer
      title:
        example: домашка
        type: string
      version:
        example: 1
        type: integer
    type: object
  internal_features_tasks_transport_http.PatchTaskRequest:
    properties:
      completed:
        type: boolean
      description:
        example: "null"
        type: string
      title:
        example: домашка
        type: string
    type: object
  internal_features_tasks_transport_http.PatchUserResponse:
    properties:
      author_user_id:
        example: 1
        type: integer
      completed:
        example: false
        type: boolean
      completed_at:
        example: "null"
        type: string
      created_at:
        example: "2026-08-26T10:30:00Z"
        type: string
      description:
        example: сделать домашку
        type: string
      id:
        example: 1
        type: integer
      title:
        example: домашка
        type: string
      version:
        example: 1
        type: integer
    type: object
  internal_features_tasks_transport_http.TaskDTOResponse:
    properties:
      author_user_id:
        example: 1
        type: integer
      completed:
        example: false
        type: boolean
      completed_at:
        example: "null"
        type: string
      created_at:
        example: "2026-08-26T10:30:00Z"
        type: string
      description:
        example: сделать домашку
        type: string
      id:
        example: 1
        type: integer
      title:
        example: домашка
        type: string
      version:
        example: 1
        type: integer
    type: object
  internal_features_users_transport_http.CreateUserRequest:
    properties:
      full_name:
        example: Petya
        maxLength: 100
        minLength: 3
        type: string
      phone_number:
        example: "+375777777777"
        maxLength: 15
        minLength: 10
        type: string
    required:
    - full_name
    type: object
  internal_features_users_transport_http.CreateUserResponse:
    properties:
      full_name:
        example: Petya
        type: string
      id:
        example: 1
        type: integer
      phone_number:
        example: "+375777777777"
        type: string
      version:
        example: 1
        type: integer
    type: object
  internal_features_users_transport_http.GetUserResponse:
    properties:
      full_name:
        example: Petya
        type: string
      id:
        example: 1
        type: integer
      phone_number:
        example: "+375777777777"
        type: string
      version:
        example: 1
        type: integer
    type: object
  internal_features_users_transport_http.PatchUserRequest:
    properties:
      full_name:
        example: Petya
        type: string
      phone_number:
        example: "+375777777777"
        type: string
    type: object
  internal_features_users_transport_http.PatchUserResponse:
    properties:
      full_name:
        example: Petya
        type: string
      id:
        example: 1
        type: integer
      phone_number:
        example: "+375777777777"
        type: string
      version:
        example: 1
        type: integer
    type: object
  internal_features_users_transport_http.UserDTORepsonse:
    properties:
      full_name:
        example: Petya
        type: string
      id:
        example: 1
        type: integer
      phone_number:
        example: "+375777777777"
        type: string
      version:
        example: 1
        type: integer
    type: object
host: 127.0.0.1:5050
info:
  contact: {}
  description: Todo Application REST-API scheme
  title: Go TodoApp API
  version: "1.0"
paths:
  /Statistics:
    get:
      description: Получение статистики с опциональной фильтрацией по user_id и/или
        временному промежутку
      parameters:
      - description: Фильтрация задач по ID автора
        in: query
        name: user_id
        type: integer
      - description: 'Начало промежутка рассмотрения статистики (включительно), формат: '
        in: query
        name: from
        type: string
      - description: 'Конец промежутка рассмотрения статистики (не включительно),
          формат: '
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешное получение статистики
          schema:
            $ref: '#/definitions/internal_features_statistics_transport_http.GetStatisticsResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_vmkzy_todoapp-go_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_vmkzy_todoapp-go_internal_core_transport_http_response.ErrorResponse'
      summary: Получение статистики
      tags:
      - statistics
  /tasks:
    get:
      description: Получение списка всех задач в системе с опциональной пагинацией
        по ID автора задачи
      parameters:
      - description: Фильтрация задач по ID автора
        in: query
        name: user_id
        type: integer
      - description: Размер страницы с задачами
        in: query
        name: limit
        type: integer
      - description: Смещение страницы с задачами
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное получение списка
          schema:
            items:
              $ref: '#/definitions/internal_features_tasks_transport_http.TaskDTOResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_vmkzy_todoapp-go_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_vmkzy_todoapp-go_internal_core_transport_http_response.ErrorResponse'
      summary: Список задач
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Создать новую задаччу
      parameters:
      - description: CreateTask тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_tasks_transport_http.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Успешно созданная задача
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.CreateTaskResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_vmkzy_todoapp-go_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Author not found
          schema:
            $ref: '#/definitions/github_com_vmkzy_todoapp-go_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_vmkzy_todoapp-go_internal_core_transport_http_response.ErrorResponse'
      summary: Создать задачу
      tags:
      - tasks
  /tasks/{id}:
    delete:
      description: Удаление существующей задачи по ID
      parameters:
      - description: ID удаляемой задачи
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Успешное удаление пользователя
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_vmkzy_todoapp-go_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/github_com_vmkzy_todoapp-go_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_vmkzy_todoapp-go_internal_core_transport_http_response.ErrorResponse'
      summary: Удаление задачи
      tags:
      - tasks
    get:
      description: Получение существующей задачи по ID
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Задача успешно найдена
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.GetTaskResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_vmkzy_todoapp-go_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/github_com_vmkzy_todoapp-go_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_vmkzy_todoapp-go_internal_core_transport_http_response.ErrorResponse'
      summary: Получение задачи
      tags:
      - tasks
    patch:
      consumes:
      - application/json
      description: |-
        Обновить существующую задачу в система
        ### Логика обновление полей (Three-state logic):
        1. **Поле не передано**: `description` игнорируется, значение в бд не меняется
        2. **Явно передано значение**: `description:"..."` устанавливается новое значение в бд
        3. **Передано null**: `description:null` очищается поле в бд
        Ограничения: `title` и `completed` не могут быть выставлен как null
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: integer
      - description: PatchTask тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_tasks_transport_http.PatchTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Задача успешно изменена
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.PatchUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_vmkzy_todoapp-go_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/github_com_vmkzy_todoapp-go_internal_core_transport_http_response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_vmkzy_todoapp-go_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_vmkzy_todoapp-go_internal_core_transport_http_response.ErrorResponse'
      summary: Обновить задачу
      tags:
      - tasks
  /users:
    get:
      description: Получение списка всех пользователей в системе с опциональной пагинацией
      parameters:
      - description: Размер страницы с пользователями
        in: query
        name: limit
        type: integer
      - description: Смещение страницы с пользователями
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное получение списка
          schema:
            items:
              $ref: '#/definitions/internal_features_users_transport_http.UserDTORepsonse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_vmkzy_todoapp-go_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_vmkzy_todoapp-go_internal_core_transport_http_response.ErrorResponse'
      summary: Список пользователей
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Создать нового пользователя
      parameters:
      - description: CreateUser тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_users_transport_http.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Успешно созданный пользователь
          schema:
            $ref: '#/definitions/internal_features_users_transport_http.CreateUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_vmkzy_todoapp-go_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_vmkzy_todoapp-go_internal_core_transport_http_response.ErrorResponse'
      summary: Создать пользователя
      tags:
      - users
  /users/{id}:
    delete:
      description: Удаление существующего пользователя по ID
      parameters:
      - description: ID удаляемого пользователя
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Успешное удаление пользователя
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_vmkzy_todoapp-go_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_vmkzy_todoapp-go_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_vmkzy_todoapp-go_internal_core_transport_http_response.ErrorResponse'
      summary: Удаление пользователя
      tags:
      - users
    get:
      description: Получение существующего пользователя по ID
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь найден
          schema:
            $ref: '#/definitions/internal_features_users_transport_http.GetUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_vmkzy_todoapp-go_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_vmkzy_todoapp-go_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_vmkzy_todoapp-go_internal_core_transport_http_response.ErrorResponse'
      summary: Получение пользователя
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: |-
        Изменение существующего пользователя по ID
        ### Логика обновление полей (Three-state logic):
        1. **Поле не передано**: `phone_bumber` игнорируется, значение в бд не меняется
        2. **Явно передано значение**: `phone_number:"+375..."` устанавливается новое значение в бд
        3. **Передано null**: `phone_number:null` очищается поле в бд
        Ограничения: `full_name` не может быть выставлен как null
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: PatchUser тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_users_transport_http.PatchUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь успешно изменен
          schema:
            $ref: '#/definitions/internal_features_users_transport_http.PatchUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_vmkzy_todoapp-go_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_vmkzy_todoapp-go_internal_core_transport_http_response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_vmkzy_todoapp-go_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_vmkzy_todoapp-go_internal_core_transport_http_response.ErrorResponse'
      summary: Изменение пользователя
      tags:
      - users
swagger: "2.0"